moved svg paths
