+++ package.json
@@ -10,12 +10,13 @@
     "start": "react-scripts start"
   },
   "dependencies": {
+    "axios": "^1.4.0",
     "framer-motion": "^8.0.2",
     "html-react-parser": "^3.0.16",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-google-recaptcha": "^2.1.0",
     "react-icons": "^4.7.1",
-    "react-pdf": "^6.2.2",
     "react-router-dom": "^6.11.2",
     "react-router-hash-link": "^2.4.3",
     "react-scripts": "^5.0.1",
